move MapContextMenue to components  folder
