+++ src/modules/map/components/contextMenue/index.js
@@ -1,0 +1,4 @@
+import { MapContextMenue } from './MapContextMenue';
+if (!window.customElements.get(MapContextMenue.tag)) {
+	window.customElements.define(MapContextMenue.tag, MapContextMenue);
+}
