+++ src/modules/map/index.js
@@ -2,5 +2,6 @@ import './i18n';
 import './components/zoomButtons';
 import './components/layerManager';
 import './components/infoButton';
+import './components/contextMenue';
 import './components/olMap';
